Fix a load of size-mismatch issues in EtoileUI on 64-bit.
Quentin: Some of these are a bit ugly because GNUstep-gui uses [unsigned] int where it should be using NS[U]Integer.  The correct fix in these cases is to fix things in -gui.  This will break the ABI, so we should really do a -gui release afterwards.



git-svn-id: http://svn.gna.org/svn/etoile/trunk/Etoile/Frameworks/EtoileUI@7918 7b3f36aa-e6db-49c1-8b4c-66eac1997e64

+++ Headers/ETController.h
@@ -139,7 +139,7 @@ objects as a collection distinct from the content. */
                           options: (NSDictionary *)options;
 - (BOOL) canMutate;
 - (BOOL) isContentMutable;
-- (unsigned int) insertionIndex;
+- (NSInteger) insertionIndex;
 - (NSIndexPath *) insertionIndexPath;
 - (NSIndexPath *) additionIndexPath;
 - (void) insertItem: (ETLayoutItem *)anItem atIndex: (NSUInteger)index;
